Fix Path Traversal Vulnerability in AutoExtract.java File Deletion Logic(Line 513)

+++ robocode.installer/src/main/java/net/sf/robocode/installer/AutoExtract.java
@@ -538,7 +538,10 @@ public class AutoExtract implements ActionListener {
                     }
                 }
             }
-            return dir.delete();
+            if (dir.exists() && !dir.isDirectory() && !dir.getAbsolutePath().contains("..")) {
+                dir.delete();
+            }
+
         }
         return false;
     }
